+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - make land command reject local changes on land
 - replace `--base` option with `--cherry-pick` in `spr diff`
+- add `--cherry-pick` option to `spr land`
 
 ## [1.2.4] - 2022-05-26